implemented path GET tradings

+++ src/main/java/org/example/app/dtos/TradeDealDTO.java
@@ -13,7 +13,7 @@ public class TradeDealDTO {
     private String id;
     private String cardToTrade;
     private String cardType;
-    private Integer minimumDamage;
+    private Double minimumDamage;
 
 }
